+++ BOCCHI.Common/Config/Fields/ShopTargetListAttribute.cs
@@ -1,4 +1,3 @@
-using BOCCHI.Common.Config;
 using BOCCHI.Common.Config.Renderers;
 using Ocelot.Config.Fields;